requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg2-binary>=2.9,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
pydantic-settings>=2.4,<3.0
email-validator>=2.2,<3.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
